+++ src/misc/lv_types.h
@@ -116,6 +116,10 @@ typedef struct _lv_theme_t lv_theme_t;
 typedef struct _lv_anim_t lv_anim_t;
 
 typedef struct _lv_font_t lv_font_t;
+typedef struct _lv_font_class_t lv_font_class_t;
+typedef struct _lv_font_info_t lv_font_info_t;
+
+typedef struct _lv_font_manager_t lv_font_manager_t;
 
 typedef struct _lv_image_decoder_t lv_image_decoder_t;
