--- src/others/font_manager/lv_font_manager_backends.h
@@ -1,69 +1,0 @@
-/**
- * @file lv_font_manager_backends.h
- *
- */
-#ifndef LV_FONT_MANAGER_BACKENDS_H
-#define LV_FONT_MANAGER_BACKENDS_H
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-/*********************
- *      INCLUDES
- *********************/
-
-#include "lv_font_manager_types.h"
-
-#if LV_USE_FONT_MANAGER
-
-/*********************
- *      DEFINES
- *********************/
-
-#define LV_BUILTIN_FONT_SRC_END {NULL, 0}
-
-/**********************
- *      TYPEDEFS
- **********************/
-
-typedef struct {
-    const lv_font_t * font_p; /**< Pointer to built-in font*/
-    uint32_t size; /** < Size of the built-in font*/
-} lv_builtin_font_src_t;
-
-LV_ATTRIBUTE_EXTERN_DATA extern const lv_font_class_t lv_builtin_font_class;
-
-#if LV_USE_FREETYPE
-/* Only path string is required */
-typedef const char lv_freetype_font_src_t;
-
-LV_ATTRIBUTE_EXTERN_DATA extern const lv_font_class_t lv_freetype_font_class;
-#endif /*LV_USE_FREETYPE*/
-
-#if LV_USE_TINY_TTF
-typedef struct {
-    const char * path; /**< Path to the font file*/
-    const void * data; /**< Pointer to the font data*/
-    size_t data_size; /**< Size of the font data*/
-    size_t cache_size; /**< Size of the font cache*/
-} lv_tiny_ttf_font_src_t;
-
-LV_ATTRIBUTE_EXTERN_DATA extern const lv_font_class_t lv_tiny_ttf_font_class;
-#endif /*LV_USE_TINY_TTF*/
-
-/**********************
- * GLOBAL PROTOTYPES
- **********************/
-
-/**********************
- *      MACROS
- **********************/
-
-#endif /*LV_USE_FONT_MANAGER*/
-
-#ifdef __cplusplus
-} /*extern "C"*/
-#endif
-
-#endif /* LV_FONT_MANAGER_BACKENDS_H */
